Run RNTR against simple product test (#42827)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/42827

This is a simple snapshot test using ReactTestRender's toJSON. We have almost the same functionality in RNTR already, so let's see if we can support tests like this.

Merging the new setup and environment with the existing configuration (https://fburl.com/code/s85sma77) still causes issues so here we add unmocking and new setup inline. Added task T177114228 to track following up on this

Note that some formatting is changed and props are dropped on the snapshot. This is resolved with refactor in next diff

Changelog: [internal]

Reviewed By: yungsters

Differential Revision: D53321823

fbshipit-source-id: a20f77c29fefb9172f8a8189bd821dd202b2ff02