Fix linking of local app modules turbomodules (#42921)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/42921

I realized that we were missing a bit to properly link app modules on Android.
The `_ModuleProvider` was never linked correctly so the TM won't be loaded at all.

With this change I'm getting the App Target (say `AppModule`) and passing it down to the
default-app-setup with a couple of macros.

This makes sure that if there is a codegen local module, we import the correct header and query the `AppModule_ModuleProvider` correctly.

Changelog:
[Android] [Fixed] - Fix linking of local app modules turbomodules

Reviewed By: cipolleschi

Differential Revision: D53567201

fbshipit-source-id: d14e61b7f2d86f15363600cd9dd1ed1ca27bd1fc