Make [RCTBridge currentBridge] work also inBridgeless for backward compatibility. (#42940)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/42940

We might have some libraries that goes through the interop layer that are using [RCTBridge currentBridge]. Although this is an API that we would like them not to use, that's the sad reality and if we don't want to force a migration on those libraries, we need to be backward compatible.

This diff sets the `RCTBridgeProxy` as the `currentBridge` in case the app is running in BridgelessMode. This should make Bridgeless backward compatible.

## Changelog
[iOS][Fixed] - Make [RCTBridge currentBridge] work in bridgeless mode

Reviewed By: RSNara

Differential Revision: D53575361

fbshipit-source-id: 179e440662b577954a577e8400e0ce0dc5b4d3ff